Split loop() code into functions. Add abstract function declarations and reorder existing functions.
This one was pretty destructive. As I broke off code into new functions, I changed some variable names in the process and reordered some of the existing functions so that they are grouped by type. The center() and reset() functions work, and I believe the polarsweep() works as well, but didn't have time to test it. These changes were needed because we are now testing multiple different commands to the mover, and it was getting messy constantly editing the loop() function.